Add files via upload
This project is a task given by Sanbercode for the final assignment of Python Data Science Course. In this project, I build machine learning to predict employee salary (above 7 million or under 7 million) base on the independent variables given. In the end, the result shows that my machine got 0,84 accuracy score out of 1.